fixing broken spec

+++ app/models/checkin.rb
@@ -45,6 +45,7 @@ class Checkin < ApplicationRecord
   end
 
   def decrement_checkin_count
+    return unless location
     location.save
     location.destroy if location.checkins_count <= 0
   end
@@ -92,6 +93,7 @@ class Checkin < ApplicationRecord
   def assign_location
     existing_location = device.locations.near([lat, lng], 0.1, units: :km).first
     location = existing_location || Location.create(lat: lat, lng: lng, device_id: device.id)
+
     update(location_id: location.id)
   end
